feat(ui): add shadow to the select menu
