+++ packages/ui/src/components/Form/Controller.tsx
@@ -14,7 +14,7 @@ import {
   PatternInputProps,
 } from "@/components/PatternInput";
 import { Rating as BaseRating } from "@/components/Rating";
-import { SelectV2 as BaseSelect, SelectProps } from "@/components/Select";
+import { Select as BaseSelect, SelectProps } from "@/components/Select";
 import { Switch as BaseSwitch, SwitchProps } from "@/components/Switch";
 import { Textarea as BaseTextarea } from "@/components/Textarea";
 import { TextareaProps } from "@/components/Textarea/Textarea";
