+++ packages/ui/src/components/Select/Select.tsx
@@ -19,104 +19,149 @@ import { isEmpty } from "lodash";
 import { SelectProps } from "@/components/Select/types";
 
 export const Select = <T extends string | number>({
+  id,
+  label,
   value,
   placeholder,
   options = [],
   showDropdownIcon = true,
   disabled = false,
+  size = "large",
+  helper,
   onChange,
 }: SelectProps<T>) => {
   const [open, setOpen] = useState<boolean>(false);
+
+  const text = useMemo(
+    () => options.find((option) => option.value === value)?.label,
+    [options, value]
+  );
+
   const onValueChange = useCallback(
     (value: string) => {
-      if (!onChange) return;
       const optionValue = options.find(
-        (option) => option.value.toString() === value
+        (option) => option.value.toString() === value.toString()
       );
-      if (!optionValue) return;
+      if (!optionValue || !onChange) return;
       onChange(optionValue.value);
     },
     [onChange, options]
   );
 
-  const label = useMemo(
-    () => options.find((option) => option.value === value)?.label,
-    [options, value]
-  );
-
   return (
     <Root
+      dir="rtl"
       open={open}
       onOpenChange={setOpen}
-      dir="rtl"
       value={value?.toString()}
       onValueChange={onValueChange}
       disabled={isEmpty(options) || disabled}
     >
-      <Trigger
-        data-open={open}
-        disabled={disabled}
-        className={cn(
-          "tw-flex tw-flex-row tw-justify-between tw-items-center",
-          "tw-w-full tw-h-14 tw-rounded-lg tw-p-2 tw-transition-colors tw-duration-200",
-          "tw-border tw-border-natural-300 hover:tw-border-brand-200 focus:tw-border-brand-500",
-          "data-[error=true]:tw-border-destructive-500 data-[error=true]:tw-shadow-ls-x-small data-[error=true]:tw-shadow-[rgba(204,0,0,0.25)]",
-          "focus:tw-outline-none focus:tw-shadow-ls-x-small focus:tw-shadow-[rgba(43,181,114,0.25)]",
-          "tw-bg-natural-50 hover:tw-bg-brand-50",
-          "data-[open=true]:tw-shadow-ls-x-small data-[open=true]:tw-shadow-[rgba(43,181,114,0.25)] data-[open=true]:tw-border-brand-500",
-          "disabled:tw-cursor-not-allowed"
-        )}
-      >
-        <Value
-          placeholder={
-            placeholder ? (
-              <Typography tag="span" className={cn("tw-text-natural-400")}>
-                {placeholder}
-              </Typography>
-            ) : null
-          }
-        >
-          <Typography tag="span" className={cn("tw-text-natural-900")}>
+      <div>
+        {label ? (
+          <Typography
+            htmlFor={id}
+            tag="label"
+            className="tw-mb-1 tw-text-natural-950 tw-text-caption tw-font-semibold"
+          >
             {label}
           </Typography>
-        </Value>
-        {showDropdownIcon ? (
-          <Icon>
-            <ArrowDown
-              data-open={open}
-              className={cn(
-                "tw-justify-self-end",
-                "data-[open=true]:tw-rotate-180 tw-transition-all tw-duration-300"
-              )}
-            />
-          </Icon>
         ) : null}
-      </Trigger>
-      <Portal>
-        <Content
-          position="popper"
+
+        <Trigger
+          id={id}
+          data-open={open}
+          disabled={disabled}
+          role="button"
           className={cn(
-            "tw-bg-natural-50 tw-border tw-border-brand-400 tw-p-1 tw-rounded-lg",
-            "tw-w-[var(--radix-select-trigger-width)] tw-z-select-dropdown"
+            "tw-flex tw-flex-row tw-justify-between tw-items-center",
+            "tw-w-full tw-rounded-lg tw-p-2",
+            "tw-bg-natural-50 tw-transition-colors tw-duration-200",
+            "tw-border tw-border-natural-300",
+            "tw-transition-colors tw-duration-200",
+            "disabled:tw-cursor-not-allowed",
+            "focus:tw-outline-none focus:tw-ring-1 focus:tw-ring-secondary-600 focus:tw-border-secondary-600",
+            {
+              "tw-h-7": size === "small",
+              "tw-h-8": size === "medium",
+              "tw-h-10": size === "large",
+            }
           )}
-          sideOffset={12}
         >
-          <Viewport className="focus:tw-outline-red-500">
-            <Group
+          <Value
+            placeholder={
+              <Typography
+                tag="span"
+                className={cn(
+                  "tw-text-natural-600 tw-font-medium tw-text-caption"
+                )}
+              >
+                {placeholder}
+              </Typography>
+            }
+          >
+            <Typography
+              tag="span"
               className={cn(
-                "tw-flex tw-flex-col tw-gap-1 tw-max-h-64 tw-overflow-y-auto",
-                "tw-scrollbar-thin tw-scrollbar-thumb-neutral-200 tw-scrollbar-track-transparent"
+                "tw-text-natural-800 tw-text-caption tw-font-meium"
               )}
             >
-              {options.map((option) => (
-                <SelectItem value={option.value.toString()} key={option.value}>
-                  {option.label}
-                </SelectItem>
-              ))}
-            </Group>
-          </Viewport>
-        </Content>
-      </Portal>
+              {text}
+            </Typography>
+          </Value>
+          {showDropdownIcon ? (
+            <Icon>
+              <ArrowDown
+                data-open={open}
+                className={cn(
+                  "tw-w-4 tw-h-4",
+                  "tw-justify-self-end",
+                  "data-[open=true]:tw-rotate-180 tw-transition-all tw-duration-300"
+                )}
+              />
+            </Icon>
+          ) : null}
+        </Trigger>
+
+        {helper ? (
+          <Typography
+            tag="span"
+            className="tw-mt-1 tw-text-natural-600 tw-text-tiny tw-font-semibold"
+          >
+            {helper}
+          </Typography>
+        ) : null}
+
+        <Portal>
+          <Content
+            position="popper"
+            className={cn(
+              "tw-bg-natural-50 tw-border tw-border-natural-200 tw-rounded-lg",
+              "tw-w-[var(--radix-select-trigger-width)] tw-z-select-dropdown tw-overflow-hidden tw-shadow-select-menu"
+            )}
+            sideOffset={helper ? 26 : 4}
+          >
+            <Viewport>
+              <Group
+                className={cn(
+                  "tw-flex tw-flex-col tw-max-h-[204px] tw-overflow-y-auto",
+                  "tw-scrollbar-thin tw-scrollbar-thumb-neutral-200 tw-scrollbar-track-transparent"
+                )}
+              >
+                {options.map((option) => (
+                  <SelectItem
+                    value={option.value.toString()}
+                    key={option.value}
+                    disabled={option.disabled}
+                  >
+                    {option.label}
+                  </SelectItem>
+                ))}
+              </Group>
+            </Viewport>
+          </Content>
+        </Portal>
+      </div>
     </Root>
   );
 };
@@ -126,19 +171,25 @@ const SelectItem = React.forwardRef<HTMLDivElement, SelectItemProps>(
     return (
       <Item
         className={cn(
-          "tw-h-14 tw-flex tw-shrink-0 tw-items-center tw-justify-between",
-          "data-[state=unchecked]:hover:tw-bg-natural-100 data-[state=unchecked]:hover:tw-text-natural-900",
-          "focus:tw-outline-none data-[state=unchecked]:focus:tw-bg-natural-100",
-          "data-[state=unchecked]:active:tw-bg-brand-700 data-[state=unchecked]:active:tw-text-natural-50",
-          "tw-cursor-pointer tw-px-3 tw-rounded-lg",
-          "data-[state=checked]:tw-bg-brand-700 data-[state=checked]:tw-text-natural-50",
+          // bg color
+          "tw-bg-natural-50 data-[state=unchecked]:focus:tw-bg-natural-100",
+          "data-[state=checked]:tw-bg-brand-500",
+          // text color
+          "tw-text-natural-600 data-[state=checked]:tw-text-natural-50 tw-rounded-lg",
+          // display and dimentions
+          "tw-flex tw-shrink-0 tw-items-center tw-justify-between",
+          "tw-h-[34px] tw-p-2",
+          "tw-cursor-pointer focus:tw-outline-none focus:tw-ring-inset focus:tw-ring-1 focus:tw-ring-secondary-600 focus:tw-z-select-item",
+          "data-[disabled]:tw-cursor-not-allowed",
           className
         )}
         {...props}
         ref={ref}
       >
-        <ItemText className="tw-text-natural-900">
-          <Typography tag="div">{children}</Typography>
+        <ItemText>
+          <Typography tag="div" className="tw-text-tiny tw-font-semibold">
+            {children}
+          </Typography>
         </ItemText>
       </Item>
     );
