+++ packages/ui/src/components/Select/index.ts
@@ -1,3 +1,2 @@
 export { Select } from "@/components/Select/Select";
-export { SelectV2 } from "@/components/Select/SelectV2";
 export type { SelectList, SelectProps } from "@/components/Select/types";
